requirements.txt
Django==2.1.5
django-debug-toolbar==1.11
django-extensions==2.1.5
djangorestframework==3.9.1
invoke==1.2.0
pkg-resources==0.0.0
pytz==2018.9
six==1.12.0
sqlparse==0.2.4
